Skip tests related to PersistentObject if that class is missing
